workaround Arduino bug in conditional includes
Flip the #ifndef conditional inclusion of date functions from:

#ifndef NODATEFUNCTIONS
. . date stuff
#endif

to:

#ifdef DATEFUNCTIONS
. . date stuff
#endif

The first way is better and this change makes it slightly inconvenient for those who
want data functions (like…uh..the original author).  But unfortunately the first form
does not work apparently due to a bug in the way the Arduino IDE/compiler handlers
pre-processor statements.

If may be a variant of: http://code.google.com/p/arduino/issues/detail?id=455

The use case and symptoms are somewhat different, but it feels like a similar problem.